Adding timer implementation together with stupid rpm algorithm setter.
This commit was tested on HW and it seems it works.

Currently setting rpm is not exact and require improvment from HW level
side.

User can set time and rpm and run the spin coater. Spin coater will try
to set best parameters to spin for given time and rpm value.